Updated to work with new Keyvault
Not sure if this PR is the best way of doing is but I noticed that the secret jenkins-x-chartmuseum silently fails after creation. By removing the properties tags from the mappings the secret got populated, but then all services' builds failed because Chartmuseum couldn't find neither the username nor the password. By adding two redundant fields I managed to get everything working, but I feel like this is a temporary fix. You might want to look at it better, thanks.